requirements.txt
langchain 
langgraph 
openai 
pyttsx3 
pillow 
moviepy==1.0.3
google-api-python-client 
google-auth 
google-auth-oauthlib 
google-auth-httplib2
langchain-community
elevenlabs